o specify the folder where uploaded files should be saved, you can define a configuration variable called UPLOAD_FOLDER in your Flask app. Here's an example of how to do that:
In this example, the UPLOAD_FOLDER configuration variable is set to the uploads subdirectory of the Flask app's root directory. You can change this to any directory on your server that you want to use for file uploads.

Note that you will also need to create the uploads directory if it does not already exist, or else the file uploads will fail. You can do this with the following code:

if not os.path.exists(app.config['UPLOAD_FOLDER']):
    os.makedirs(app.config['UPLOAD_FOLDER'])

You can add this code to your Flask app setup to create the uploads directory automatically if it does not already exist.