example of the save_file_for_user() function that takes a description argument to save with the file:
In this example, the File model has a new description column that stores a text description of the uploaded file. The save_file_for_user() function takes a new description argument and uses it to create a new File object with the user, file path, and description. The rest of the function is the same as before.

You will also need to update your database schema and migrations to add the new description column to the File model. Depending on your ORM and migration tool, the exact process for doing this may vary.