he regular expression \.([A-Za-z0-9]+)$ matches the last dot and any sequence of alphanumeric characters at the end of the filename (i.e., the file extension).
 the regular expression \.([A-Za-z0-9]+)$ matches the last dot and any sequence of alphanumeric characters at the end of the filename (i.e., the file extension). The group(1) method returns the matched sequence of characters, which is then converted to lowercase and checked against the ALLOWED_EXTENSIONS set.

This method provides more robust validation of the file extension compared to the previous example, since it ensures that the extension is valid and doesn't contain any malicious characters or other unexpected characters. However, it may be slightly slower than the previous method, since it involves running a regular expression on each filename.