ion uses Materialize, a popular front-end framework that implements Google's Material Design guidelines.
It includes a card element that contains the form for uploading a file. The file input field is styled as a button, with a separate input field for the file path. The description input field is styled as a text field with a label. The submit button is styled with a blue color scheme and an icon on the right side.

Note that this version also includes a link to the Materialize CSS file, which needs to be included in the head section of the HTML file for the styles to be applied.