Protect views with login_require
n this example, the @login_required decorator is used to protect the /dashboard view. If a user tries to access this view without being authenticated, they will be redirected to the login page. If they are authenticated, the view function will be executed and the dashboard.html template will be rendered.