matching 3D models and text descriptions using SQLalchemy:
n this example, the search_models function takes a query parameter which is the user's search query. The function creates a SQLalchemy session to query the database and uses the query method to search for 3D models and text descriptions that match the query. The or_ function is used to search for models where either the name or description matches the query.

The results are then converted to a list of dictionaries using a to_dict method on the Model class, which is assumed to be defined in a separate models.py module. This method converts the model object to a dictionary with keys for each attribute of the model.

Finally, the function returns the list of results as a list of dictionaries.